Setup the group views. Rename the accounts URLs
It is necessary to rename the account URLs. User was implicit, and now
that there are groups, we have to rename the user views.

The views have been spit into different modules too.
