+++ server/accounts/urls.py
@@ -2,28 +2,39 @@ from django.urls import path
 
 from . import views
 
-app_name = "users"
+app_name = "accounts"
 urlpatterns = [
-    path('', views.UsersView.as_view(),
-         name="list"),
     path('nginx/auth_request/', views.NginxAuthRequestView.as_view(),
          name="nginx_auth_request"),
-    path('invite/', views.InviteUserView.as_view(),
-         name="invite"),
-    path('create_service_account/', views.CreateServiceAccountView.as_view(),
+
+    # manage users
+    path('users/', views.UsersView.as_view(),
+         name="users"),
+    path('users/invite/', views.InviteUserView.as_view(),
+         name="invite_user"),
+    path('users/create_service_account/', views.CreateServiceAccountView.as_view(),
          name="create_service_account"),
-    path('<int:pk>/', views.UserView.as_view(),
+    path('users/<int:pk>/', views.UserView.as_view(),
          name="user"),
-    path('<int:pk>/update/', views.UpdateUserView.as_view(),
-         name="update"),
-    path('<int:pk>/delete/', views.DeleteUserView.as_view(),
-         name="delete"),
-    path('<int:pk>/api_token/create/', views.CreateUserAPITokenView.as_view(),
+    path('users/<int:pk>/update/', views.UpdateUserView.as_view(),
+         name="update_user"),
+    path('users/<int:pk>/delete/', views.DeleteUserView.as_view(),
+         name="delete_user"),
+    path('users/<int:pk>/api_token/create/', views.CreateUserAPITokenView.as_view(),
          name="create_user_api_token"),
-    path('<int:pk>/api_token/view/', views.UserAPITokenView.as_view(),
+    path('users/<int:pk>/api_token/view/', views.UserAPITokenView.as_view(),
          name="user_api_token"),
-    path('<int:pk>/api_token/delete/', views.DeleteUserAPITokenView.as_view(),
+    path('users/<int:pk>/api_token/delete/', views.DeleteUserAPITokenView.as_view(),
          name="delete_user_api_token"),
+
+    # manage groups
+    path('groups/', views.GroupsView.as_view(), name="groups"),
+    path('groups/create/', views.CreateGroupView.as_view(), name="create_group"),
+    path('groups/<int:pk>/', views.GroupView.as_view(), name="group"),
+    path('groups/<int:pk>/update/', views.UpdateGroupView.as_view(), name="update_group"),
+    path('groups/<int:pk>/delete/', views.DeleteGroupView.as_view(), name="delete_group"),
+
+    # user views
     path('verification_devices/', views.UserVerificationDevicesView.as_view(),
          name="verification_devices"),
     path('add_totp/', views.AddTOTPView.as_view(),
