+++ server/accounts/views/__init__.py
@@ -1,0 +1,4 @@
+from .auth import *  # NOQA
+from .groups import *  # NOQA
+from .user import *  # NOQA
+from .users import *  # NOQA
