+++ server/templates/accounts/add_totp.html
@@ -14,7 +14,7 @@
   </div>
   {{ form.verification_code|bootstrap }}
   <p>
-    <a href="{% url 'users:verification_devices' %}" class="btn btn-default">Cancel</a>
+    <a href="{% url 'accounts:verification_devices' %}" class="btn btn-default">Cancel</a>
     <button type="submit" class="btn btn-primary">
       Save
     </button>
