+++ server/templates/accounts/delete_user.html
@@ -6,7 +6,7 @@
 <p>Please confirm you want to delete the {{ user.get_type_display }}: {{ user }}.</p>
 <form class="form" method="POST">{% csrf_token %}
   <p>
-    <a href="{% url 'users:list' %}" class="btn btn-default">Cancel</a>
+    <a href="{% url 'accounts:users' %}" class="btn btn-default">Cancel</a>
     <button type="submit" class="btn btn-danger">Delete {{ user.get_type_display }}</button>
   </p>
 </form>
