+++ server/templates/accounts/group_form.html
@@ -1,0 +1,31 @@
+{% extends 'base.html' %}
+{% load bootstrap %}
+
+{% block content %}
+<ol class="breadcrumb">
+  <li><a href="/">Home</a></li>
+  <li><a href="{% url 'accounts:groups' %}">Groups</a></li>
+  {% if object %}
+  <li><a href="{% url 'accounts:group' object.pk %}">{{ object }}</a></li>
+  <li class="active">Update</li>
+  {% else %}
+  <li class="active">Create</li>
+  {% endif %}
+</ol>
+
+<h3>{{ title }}</h3>
+
+<form class="form" method="POST">{% csrf_token %}
+  {{ form|bootstrap }}
+  <p>
+    {% if object %}
+    <a href="{% url 'accounts:group' group.pk %}" class="btn btn-default">Cancel</a>
+    {% else %}
+    <a href="{% url 'accounts:groups' %}" class="btn btn-default">Cancel</a>
+    {% endif %}
+    <button type="submit" class="btn btn-primary">
+      Save
+    </button>
+  </p>
+</form>
+{% endblock %}
