+++ server/templates/accounts/service_account_form.html
@@ -7,7 +7,7 @@
 <form class="form" method="POST">{% csrf_token %}
   {{ form|bootstrap }}
   <p>
-    <a href="{% url 'users:list' %}" class="btn btn-default">Cancel</a>
+    <a href="{% url 'accounts:users' %}" class="btn btn-default">Cancel</a>
     <button type="submit" class="btn btn-primary">
       Save
     </button>
