+++ server/templates/accounts/user_form.html
@@ -4,8 +4,8 @@
 {% block content %}
 <ol class="breadcrumb">
   <li><a href="/">Home</a></li>
-  <li><a href="{% url 'users:list' %}">Users</a></li>
-  <li><a href="{% url 'users:user' user.pk %}">{{ user.get_type_display|title }} {{ user }}</a></li>
+  <li><a href="{% url 'accounts:users' %}">Users</a></li>
+  <li><a href="{% url 'accounts:user' user.pk %}">{{ user.get_type_display|title }} {{ user }}</a></li>
   <li class="active">Update</li>
 </ol>
 
@@ -14,7 +14,7 @@
 <form class="form" method="POST">{% csrf_token %}
   {{ form|bootstrap }}
   <p>
-    <a href="{% url 'users:user' user.pk %}" class="btn btn-default">Cancel</a>
+    <a href="{% url 'accounts:user' user.pk %}" class="btn btn-default">Cancel</a>
     <button type="submit" class="btn btn-primary">
       Save
     </button>
