+++ server/templates/accounts/user_verification_devices.html
@@ -11,8 +11,8 @@
     <span class="caret"></span>
   </button>
   <ul class="dropdown-menu" aria-labelledby="addVD">
-    <li><a href="{% url 'users:add_totp' %}">TOTP</a></li>
-    <li><a href="{% url 'users:register_u2f_device' %}">U2F</a></li>
+    <li><a href="{% url 'accounts:add_totp' %}">TOTP</a></li>
+    <li><a href="{% url 'accounts:register_u2f_device' %}">U2F</a></li>
   </ul>
 </div>
